backend/requirements.txt
Flask==2.1.3
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
sqlalchemy==1.4.27
requests==2.26.0
python-dotenv==0.19.2
